# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BorgLayer Marketplace API
  version: "1.0.0"
  description: |
    Programmable domain commerce. Search, evaluate, and buy domains via REST,
    or connect an AI agent through the MCP server at mcp.borglayer.com/sse.
  contact:
    name: BorgLayer
    url: https://borglayer.com
servers:
  - url: https://api.borglayer.com/api/v1
    description: Production
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      bearerFormat: "blk_live_... / blk_test_..."
...
security:
  - ApiKey: []
paths:
...
  /checkout:
    post:
      summary: Create a checkout (escrow) for a listing
      description: Requires scope checkout.write. Records commission attribution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id, buyer_wallet]
              properties:
                listing_id: { type: string, format: uuid }
                buyer_wallet: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
      responses: { "201": { description: Checkout created with payment details } }
  /orders/{id}:
    get:
      summary: Order details
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Order } }
  /orders/{id}/status:
    get:
      summary: Order status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Status } }
...